add separate registration
