+++ lib/screens/registration.dart
@@ -2,11 +2,11 @@
 
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:victu/objects/userData.dart';
+import 'package:victu/objects/user_data.dart';
+import 'package:victu/screens/registration/canteen_registration.dart';
 import 'package:victu/screens/home_page.dart';
-import 'package:victu/screens/login.dart';
-import 'package:victu/screens/userRegistration.dart';
+import 'package:victu/screens/registration/farmer_registration.dart';
+import 'package:victu/screens/registration/user_registration.dart';
 import 'package:victu/utils/database.dart';
 
 class Registration extends StatefulWidget {
@@ -64,24 +64,24 @@ class _RegistrationState extends State<Registration> {
     );
   }
 
-  Route _routeToSignInScreen() {
-    return PageRouteBuilder(
-      pageBuilder: (context, animation, secondaryAnimation) => const Login(),
-      transitionsBuilder: (context, animation, secondaryAnimation, child) {
-        var begin = const Offset(-1.0, 0.0);
-        var end = Offset.zero;
-        var curve = Curves.ease;
+  // Route _routeToSignInScreen() {
+  //   return PageRouteBuilder(
+  //     pageBuilder: (context, animation, secondaryAnimation) => const Login(),
+  //     transitionsBuilder: (context, animation, secondaryAnimation, child) {
+  //       var begin = const Offset(-1.0, 0.0);
+  //       var end = Offset.zero;
+  //       var curve = Curves.ease;
 
-        var tween =
-            Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
+  //       var tween =
+  //           Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
 
-        return SlideTransition(
-          position: animation.drive(tween),
-          child: child,
-        );
-      },
-    );
-  }
+  //       return SlideTransition(
+  //         position: animation.drive(tween),
+  //         child: child,
+  //       );
+  //     },
+  //   );
+  // }
 
   @override
   Widget build(BuildContext context) {
@@ -210,334 +210,25 @@ class _RegistrationState extends State<Registration> {
                   ],
                 ))),
         Step(
-          isActive: currentStep >= 1,
-          title: const Text("User Registration"),
-          content: userType == UserType.student
-              ? UserRegistration(
-                  nameController: nameController,
-                  emailController: emailController,
-                  ageController: ageController,
-                  heightController: heightController,
-                  weightController: weightController)
-              : userType == UserType.canteen
-                  ? CanteenRegistration(
-                      nameController: nameController,
-                      emailController: emailController,
-                      contactNumberController: contactNumberController,
-                      canteenNameController: canteenNameController)
-                  : const Text("Farmer"),
-        )
+            isActive: currentStep >= 1,
+            title: const Text("User Registration"),
+            content: userType == UserType.student
+                ? UserRegistration(
+                    nameController: nameController,
+                    emailController: emailController,
+                    ageController: ageController,
+                    heightController: heightController,
+                    weightController: weightController)
+                : userType == UserType.canteen
+                    ? CanteenRegistration(
+                        nameController: nameController,
+                        emailController: emailController,
+                        contactNumberController: contactNumberController,
+                        canteenNameController: canteenNameController)
+                    : FarmerRegistration(
+                        nameController: nameController,
+                        emailController: emailController))
       ];
-}
-
-class CanteenRegistration extends StatefulWidget {
-  const CanteenRegistration(
-      {super.key,
-      required this.nameController,
-      required this.emailController,
-      required this.contactNumberController,
-      required this.canteenNameController});
-
-  final TextEditingController nameController;
-  final TextEditingController emailController;
-  final TextEditingController contactNumberController;
-  final TextEditingController canteenNameController;
-
-  @override
-  State<CanteenRegistration> createState() => _CanteenRegistrationState();
-}
-
-class _CanteenRegistrationState extends State<CanteenRegistration> {
-  var currentSelectedValue;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(16, 30, 16, 16),
-      child: SingleChildScrollView(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.start,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          mainAxisSize: MainAxisSize.max,
-          children: [
-            ///***If you have exported images you must have to copy those images in assets/images directory.
-            const Image(
-              image: NetworkImage(
-                  "https://cdn4.iconfinder.com/data/icons/security-overcolor/512/password_code-256.png"),
-              height: 100,
-              width: 100,
-              fit: BoxFit.cover,
-            ),
-            const Padding(
-              padding: EdgeInsets.fromLTRB(0, 16, 0, 0),
-              child: Text(
-                "Let's Get Started!",
-                textAlign: TextAlign.start,
-                overflow: TextOverflow.clip,
-                style: TextStyle(
-                  fontWeight: FontWeight.w700,
-                  fontStyle: FontStyle.normal,
-                  fontSize: 22,
-                  color: Color(0xff2d9871),
-                ),
-              ),
-            ),
-            const Padding(
-              padding: EdgeInsets.fromLTRB(50, 8, 50, 0),
-              child: Text(
-                "Provide your business details so we can get you all set up.",
-                textAlign: TextAlign.center,
-                overflow: TextOverflow.clip,
-                style: TextStyle(
-                  fontWeight: FontWeight.w400,
-                  fontStyle: FontStyle.normal,
-                  fontSize: 14,
-                  color: Color(0xff000000),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(0, 40, 0, 0),
-              child: TextField(
-                enabled: false,
-                controller: widget.nameController,
-                obscureText: false,
-                textAlign: TextAlign.start,
-                maxLines: 1,
-                style: const TextStyle(
-                  fontWeight: FontWeight.w400,
-                  fontStyle: FontStyle.normal,
-                  fontSize: 14,
-                  color: Color(0xff000000),
-                ),
-                decoration: InputDecoration(
-                  disabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  focusedBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  enabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  hintText: "Name",
-                  hintStyle: const TextStyle(
-                    fontWeight: FontWeight.w400,
-                    fontStyle: FontStyle.normal,
-                    fontSize: 14,
-                    color: Color(0xff000000),
-                  ),
-                  filled: true,
-                  fillColor: const Color(0xffebebeb),
-                  isDense: false,
-                  contentPadding:
-                      const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-                  prefixIcon: const Icon(Icons.person,
-                      color: Color(0xff212435), size: 24),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(0, 16, 0, 0),
-              child: TextField(
-                enabled: false,
-                controller: widget.emailController,
-                obscureText: false,
-                textAlign: TextAlign.start,
-                maxLines: 1,
-                style: const TextStyle(
-                  fontWeight: FontWeight.w400,
-                  fontStyle: FontStyle.normal,
-                  fontSize: 14,
-                  color: Color(0xff000000),
-                ),
-                decoration: InputDecoration(
-                  disabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  focusedBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  enabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  hintText: "Email Address",
-                  hintStyle: const TextStyle(
-                    fontWeight: FontWeight.w400,
-                    fontStyle: FontStyle.normal,
-                    fontSize: 14,
-                    color: Color(0xff000000),
-                  ),
-                  filled: true,
-                  fillColor: const Color(0xffebebeb),
-                  isDense: false,
-                  contentPadding:
-                      const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-                  prefixIcon: const Icon(Icons.mail,
-                      color: Color(0xff212435), size: 24),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(0, 16, 0, 0),
-              child: TextField(
-                controller: widget.canteenNameController,
-                obscureText: false,
-                textAlign: TextAlign.start,
-                maxLines: 1,
-                style: const TextStyle(
-                  fontWeight: FontWeight.w400,
-                  fontStyle: FontStyle.normal,
-                  fontSize: 14,
-                  color: Color(0xff000000),
-                ),
-                decoration: InputDecoration(
-                  disabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  focusedBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  enabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  hintText: "Canteen Name",
-                  hintStyle: const TextStyle(
-                    fontWeight: FontWeight.w400,
-                    fontStyle: FontStyle.normal,
-                    fontSize: 14,
-                    color: Color(0xff000000),
-                  ),
-                  filled: true,
-                  fillColor: const Color(0xffffffff),
-                  isDense: false,
-                  contentPadding:
-                      const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-                  prefixIcon: const Icon(Icons.shopping_cart,
-                      color: Color(0xff212435), size: 24),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(0, 16, 0, 0),
-              child: TextField(
-                keyboardType: TextInputType.number,
-                inputFormatters: <TextInputFormatter>[
-                  FilteringTextInputFormatter.digitsOnly
-                ],
-                controller: widget.contactNumberController,
-                obscureText: false,
-                textAlign: TextAlign.start,
-                maxLines: 1,
-                style: const TextStyle(
-                  fontWeight: FontWeight.w400,
-                  fontStyle: FontStyle.normal,
-                  fontSize: 14,
-                  color: Color(0xff000000),
-                ),
-                decoration: InputDecoration(
-                  disabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  focusedBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  enabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8.0),
-                    borderSide:
-                        const BorderSide(color: Color(0xff2d9871), width: 1),
-                  ),
-                  hintText: "Contact number (09xxxxxxxxx)",
-                  hintStyle: const TextStyle(
-                    fontWeight: FontWeight.w400,
-                    fontStyle: FontStyle.normal,
-                    fontSize: 14,
-                    color: Color(0xff000000),
-                  ),
-                  filled: true,
-                  fillColor: const Color(0xffffffff),
-                  isDense: false,
-                  contentPadding:
-                      const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
-                  prefixIcon: const Icon(Icons.phone,
-                      color: Color(0xff212435), size: 24),
-                ),
-              ),
-            ),
-            Row(children: [
-              Expanded(
-                  flex: 1,
-                  child: Padding(
-                    padding: const EdgeInsets.fromLTRB(0, 15, 0, 0),
-                    child: Container(
-                        width: 130,
-                        height: 50,
-                        padding: const EdgeInsets.symmetric(
-                            vertical: 4, horizontal: 13),
-                        decoration: BoxDecoration(
-                          color: const Color(0xffffffff),
-                          border: Border.all(
-                              color: const Color(0xff2d9871), width: 1),
-                          borderRadius: BorderRadius.circular(8),
-                        ),
-                        child: DropdownButtonHideUnderline(
-                          child: DropdownButton(
-                            hint: const Text("School"),
-                            value: currentSelectedValue,
-                            items: ["Sample School 1", "Sample School 2"]
-                                .map<DropdownMenuItem<String>>((String value) {
-                              return DropdownMenuItem<String>(
-                                value: value,
-                                child: Text(value),
-                              );
-                            }).toList(),
-                            style: const TextStyle(
-                              color: Color(0xff000000),
-                              fontSize: 16,
-                              fontWeight: FontWeight.w400,
-                              fontStyle: FontStyle.normal,
-                            ),
-                            onChanged: (newValue) {
-                              setState(() {
-                                currentSelectedValue = newValue;
-                              });
-                            },
-                            icon: const Icon(Icons.account_balance),
-                            iconSize: 24,
-                            iconEnabledColor: const Color(0xff212435),
-                            elevation: 8,
-                            isExpanded: true,
-                          ),
-                        )),
-                  ))
-            ]),
-          ],
-        ),
-      ),
-    );
-  }
 }
 
 enum UserType { student, canteen, farmer }
